Iteration #5

+++ app/src/main/java/edu/umkc/mes6ybmail/thedish/RecipeGateway.java
@@ -105,7 +105,7 @@ public class RecipeGateway {
         assert x == 1;
     }
 
-    public boolean delete(int id) { // id is primary key for record
+    public boolean delete(long id) { // id is primary key for record
         DatabaseHelper dbHelper = new DatabaseHelper(context);
         SQLiteDatabase db = dbHelper.getWritableDatabase();
